requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.9.5
aiohttp-retry==2.9.1
aiosignal==1.3.1
alembic==1.13.2
altair==6.0.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
apache-airflow==2.9.3
apache-airflow-providers-common-io==1.3.2
apache-airflow-providers-common-sql==1.14.2
apache-airflow-providers-fab==1.2.2
apache-airflow-providers-ftp==3.10.0
apache-airflow-providers-http==4.12.0
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-smtp==1.7.1
apache-airflow-providers-sqlite==3.8.1
apispec==6.6.1
appdirs==1.4.4
argcomplete==3.4.0
asgiref==3.8.1
astunparse==1.6.3
asyncssh==2.22.0
atpublic==7.0.0
attrs==23.2.0
Babel==2.15.0
billiard==4.2.4
bleach==6.3.0
blinker==1.8.2
cachelib==0.9.0
cachetools==7.0.5
celery==5.6.3
certifi==2024.7.4
cffi==2.0.0
charset-normalizer==3.3.2
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
clickclick==20.10.2
cloudpickle==3.1.2
colorama==0.4.6
colorlog==4.8.0
configobj==5.0.9
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.3.3
cron-descriptor==1.4.3
croniter==2.0.5
cryptography==41.0.7
cycler==0.12.1
databricks-sdk==0.102.0
Deprecated==1.2.14
dictdiffer==0.9.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docker==7.1.0
docutils==0.16
dpath==2.2.0
dulwich==1.1.0
dvc==3.67.1
dvc-data==3.18.3
dvc-http==2.32.0
dvc-objects==5.2.0
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
email_validator==2.2.0
entrypoints==0.4
fastapi==0.135.3
filelock==3.25.2
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.3.0
flask-cors==6.0.2
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flatbuffers==25.12.19
flatten-dict==0.4.2
flufl.lock==9.0.0
fonttools==4.62.1
frozenlist==1.4.1
fsspec==2026.3.0
funcy==2.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
google-pasta==0.2.0
google-re2==1.1.20240702
googleapis-common-protos==1.63.2
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenlet==3.0.3
grpcio==1.64.1
gto==1.9.0
gunicorn==22.0.0
h11==0.14.0
h5py==3.14.0
httpcore==1.0.5
httpx==0.27.0
huey==2.6.0
hydra-core==1.3.2
idna==3.7
importlib-metadata==6.11.0
importlib_resources==6.4.0
inflection==0.5.1
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.5.3
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kaggle==2.0.1
kagglesdk==0.1.18
keras==3.14.0
kiwisolver==1.5.0
kombu==5.6.2
lazy-object-proxy==1.10.0
libclang==18.1.1
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib==3.10.8
mdit-py-plugins==0.4.1
mdurl==0.1.2
methodtools==0.4.7
ml_dtypes==0.5.4
mlflow==3.11.1
mlflow-skinny==3.11.1
mlflow-tracing==3.11.1
more-itertools==10.3.0
multidict==6.0.5
namex==0.1.0
narwhals==2.19.0
networkx==3.6.1
numpy==2.4.4
omegaconf==2.3.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opt_einsum==3.4.0
optree==0.19.0
ordered-set==4.1.0
orjson==3.11.8
packaging==24.1
pandas==2.3.3
pathspec==0.12.1
pendulum==3.0.0
pillow==12.2.0
platformdirs==3.11.0
pluggy==1.5.0
prettytable==3.17.0
prison==0.2.1
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==7.34.1
psutil==6.0.0
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
pydot==4.0.1
pygit2==1.19.2
Pygments==2.18.0
pygtrie==2.5.0
PyJWT==2.8.0
pyparsing==3.3.2
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.24
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.7.1
rich-argparse==1.5.2
rpds-py==0.19.0
ruamel.yaml==0.19.1
scikit-learn==1.8.0
scipy==1.17.1
scmrepo==3.6.2
semver==3.0.4
setproctitle==1.3.3
setuptools==66.1.1
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
six==1.16.0
skops==0.13.0
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.0
sqltrie==0.11.2
starlette==1.0.0
streamlit==1.56.0
tabulate==0.9.0
tenacity==8.5.0
tensorflow_cpu==2.21.0
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.6.0
time-machine==2.14.2
toml==0.10.2
tomlkit==0.14.0
tornado==6.5.5
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.2
urllib3==2.6.3
uvicorn==0.44.0
vine==5.1.0
voluptuous==0.16.0
watchdog==6.0.0
wcwidth==0.6.0
webencodings==0.5.1
Werkzeug==2.2.3
wheel==0.46.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zc.lockfile==4.0
zipp==3.19.2
